feat: Flutter UI redesign v1.5.0 with fixed compile errors Co-authored-by: monkeycode-ai <monkeycode-ai@chaitin.com>

+++ dsh_android/lib/views/settings_view.dart
@@ -73,7 +73,7 @@ class _SettingsViewState extends ConsumerState<SettingsView> {
                 max: 120,
                 divisions: 23,
                 labelSuffix: 's',
-                onChanged: (v) => setState(() => _toolTimeoutMs = v * 1000),
+                onChanged: (v) => setState(() => _toolTimeoutMs = (v * 1000).toInt()),
               ),
             ],
           ),
